refactor(sync): relocate cloud sync files under sandbox-cloud/src/sync/ (R1)
Co-location: all cloud sync operations now live under sync/. Pure file moves +
import repointing, behaviour-preserving (no logic changes). External packages
import via the package index only, so the public API is unchanged.

Moved into sandbox-cloud/src/sync/: sync-transport, agent-credentials,
claude-json-overlay, codex-agents-override, dynamic-sync, git-identity, env-files,
carry, workspace-seed, workspace-resync (cloud-sync already there). Repointed
cloud-provider/bootstrap-launch/checkpoint/cloud-cp/index + the moved files'
./shell import + cloud-sync's imports + the test imports/vi.mock paths.

Gate: build + typecheck + lint green; cloud 99 tests green.
