+++ packages/sandbox-cloud/src/sync/claude-json-overlay.ts
